introduce the distinction between 'native' and 'remote' flavours.
always apply the basic 'native' flavour, which in turn doesn't need a setup method anymore, since it only performs basic bootstrapping (mainly configuring sshd and adding the admin user). instead the flavour supplied on the command line is always inerpreted as a ezjailremote flavour. also, we provide the ability to apply such a 'remote' flavour to existing jails via the apply_flavour task. (which is also what we call at the end of jail creation, after it has started up.)